+++ demo/src/main/java/com/orange/ods/demo/ui/components/bottomnavigation/ComponentBottomNavigation.kt
@@ -11,14 +11,9 @@
 package com.orange.ods.demo.ui.components.bottomnavigation
 
 import androidx.annotation.DrawableRes
-import androidx.annotation.StringRes
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.material.Icon
-import androidx.compose.material.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableStateOf
@@ -33,53 +28,40 @@ import com.orange.ods.compose.component.bottomnavigation.OdsBottomNavigation
 import com.orange.ods.compose.component.bottomnavigation.OdsBottomNavigationItem
 import com.orange.ods.demo.R
 import com.orange.ods.demo.ui.components.Component
-import com.orange.ods.demo.ui.components.utilities.ComponentHeader
+import com.orange.ods.demo.ui.components.ComponentDetail
+import com.orange.ods.demo.ui.components.bottomnavigation.ComponentBottomNavigation.MaxNavigationItemCount
+import com.orange.ods.demo.ui.components.bottomnavigation.ComponentBottomNavigation.MinNavigationItemCount
+import com.orange.ods.demo.ui.components.utilities.ComponentCountRow
 import com.orange.ods.demo.ui.components.utilities.clickOnElement
-import com.orange.ods.demo.ui.utilities.composable.LabelledRadioButton
 
-private data class BottomNavigationVariant(@StringRes val titleRes: Int, val itemNumber: Int)
-
-private val bottomNavigationVariants = listOf(
-    BottomNavigationVariant(R.string.component_bottom_navigation_three_up, 3),
-    BottomNavigationVariant(R.string.component_bottom_navigation_four_up, 4),
-    BottomNavigationVariant(R.string.component_bottom_navigation_five_up, 5)
-)
-
-@Composable
-fun ComponentBottomNavigation() {
-    val selectedVariantIndex = rememberSaveable { mutableStateOf(0) }
-
-    Scaffold(
-        bottomBar = { BottomNavigationBar(selectedVariantIndex) }
-    ) { innerPadding ->
-        Box(modifier = Modifier.padding(innerPadding)) {
-            Column(
-                modifier = Modifier.verticalScroll(rememberScrollState())
-            ) {
-                ComponentHeader(
-                    imageRes = R.drawable.il_bottom_navigation,
-                    imageAlignment = Component.BottomNavigation.imageAlignment,
-                    description = R.string.component_bottom_navigation_description
-                )
-                Column(
-                    modifier = Modifier.padding(top = dimensionResource(id = R.dimen.spacing_m))
-                ) {
-                    bottomNavigationVariants.forEachIndexed { index, bottomNavigationVariant ->
-                        LabelledRadioButton(
-                            selectedRadio = mutableStateOf(bottomNavigationVariants[selectedVariantIndex.value].titleRes.toString()),
-                            currentRadio = bottomNavigationVariant.titleRes.toString(),
-                            label = stringResource(id = bottomNavigationVariant.titleRes),
-                            onClick = { selectedVariantIndex.value = index }
-                        )
-                    }
-                }
-            }
-        }
-    }
+private object ComponentBottomNavigation {
+    const val MinNavigationItemCount = 3
+    const val MaxNavigationItemCount = 5
 }
 
 @Composable
-private fun BottomNavigationBar(selectedVariantIndex: MutableState<Int>) {
+@ExperimentalMaterialApi
+fun ComponentBottomNavigation() {
+    val selectedNavigationItemCount = rememberSaveable { mutableStateOf(MinNavigationItemCount) }
+    ComponentDetail(
+        component = Component.BottomNavigation,
+        bottomBar = { ComponentBottomNavigationBottomBar(selectedNavigationItemCount = selectedNavigationItemCount) },
+        content = { ComponentBottomNavigationContent(selectedNavigationItemCount = selectedNavigationItemCount) })
+}
+
+@Composable
+private fun ComponentBottomNavigationContent(selectedNavigationItemCount: MutableState<Int>) {
+    ComponentCountRow(
+        modifier = Modifier.padding(horizontal = dimensionResource(id = R.dimen.ods_screen_horizontal_margin)),
+        title = stringResource(id = R.string.component_bottom_navigation_navigation_item_count),
+        count = selectedNavigationItemCount,
+        minCount = MinNavigationItemCount,
+        maxCount = MaxNavigationItemCount
+    )
+}
+
+@Composable
+private fun ComponentBottomNavigationBottomBar(selectedNavigationItemCount: MutableState<Int>) {
     val context = LocalContext.current
     val navigationItems = listOf(
         NavigationItem("Favorites", R.drawable.ic_heart),
@@ -89,20 +71,21 @@ private fun BottomNavigationBar(selectedVariantIndex: MutableState<Int>) {
         NavigationItem("Settings", R.drawable.ic_settings)
     )
 
-    val selectedItem = remember { mutableStateOf(navigationItems[0]) }
+    val selectedNavigationItem = remember { mutableStateOf(navigationItems[0]) }
 
     OdsBottomNavigation {
-        for (item in navigationItems.subList(0, bottomNavigationVariants[selectedVariantIndex.value].itemNumber)) {
-            OdsBottomNavigationItem(
-                icon = { Icon(painter = painterResource(id = item.icon), contentDescription = null) },
-                label = item.title,
-                selected = selectedItem.value.title == item.title,
-                onClick = {
-                    selectedItem.value = item
-                    clickOnElement(context, item.title)
-                }
-            )
-        }
+        navigationItems.take(selectedNavigationItemCount.value)
+            .forEach { navigationItem ->
+                OdsBottomNavigationItem(
+                    icon = { Icon(painter = painterResource(id = navigationItem.icon), contentDescription = null) },
+                    label = navigationItem.title,
+                    selected = selectedNavigationItem.value.title == navigationItem.title,
+                    onClick = {
+                        selectedNavigationItem.value = navigationItem
+                        clickOnElement(context, navigationItem.title)
+                    }
+                )
+            }
     }
 }